Better logic.

+++ raspberry/cam-amusing.sh
@@ -51,26 +51,29 @@ for i in $(echo '01 02 03 04 05 06 07 08 09 10'); do
 	/usr/bin/streamer -c /dev/video-cam0 -r 2 -s 800x600 -o $PREFIX2-$i.jpeg 2>/dev/null
 done
 
-rm $RAMDISK/img/*-{01..09}.jpeg
+rm $RAMDISK/img/*-{01..09}.jpeg 2>/dev/null
 
-if [ -f "$RAMDISK/img/cam1.jpeg" -a -f "$RAMDISK/img/cam2.jpeg" ]; then
-	VALUE1=$(compare $RAMDISK/img/cam1.jpeg $PREFIX1-10.jpeg)
-	VALUE2=$(compare $RAMDISK/img/cam2.jpeg $PREFIX2-10.jpeg)
+if [ -f "$PREFIX1-10.jpeg" -a -f "$PREFIX2-10.jpeg"]; then
+
+	echo "archa_box2_cam1;phototrapimg;$(base64 -w0 $PREFIX1-10.jpeg);$ISO" | /bin/gzip > $PREFIX1.csv.gz
+	echo "archa_box2_cam2;phototrapimg;$(base64 -w0 $PREFIX2-10.jpeg);$ISO" | /bin/gzip > $PREFIX2.csv.gz
+
+	if [ -f "$RAMDISK/img/cam1.jpeg" -a -f "$RAMDISK/img/cam2.jpeg" ]; then
+		VALUE1=$(compare $RAMDISK/img/cam1.jpeg $PREFIX1-10.jpeg)
+		VALUE2=$(compare $RAMDISK/img/cam2.jpeg $PREFIX2-10.jpeg)
+	fi
+
+	mv $PREFIX1-10.jpeg $RAMDISK/img/cam1.jpeg
+	mv $PREFIX2-10.jpeg $RAMDISK/img/cam2.jpeg
 fi
-
-mv $PREFIX1-10.jpeg $RAMDISK/img/cam1.jpeg
-mv $PREFIX2-10.jpeg $RAMDISK/img/cam2.jpeg
-
-echo "archa_box2_cam1;phototrapimg;$(base64 -w0 $RAMDISK/img/cam1.jpeg);$ISO" | /bin/gzip > $PREFIX1.csv.gz
-echo "archa_box2_cam2;phototrapimg;$(base64 -w0 $RAMDISK/img/cam2.jpeg);$ISO" | /bin/gzip > $PREFIX2.csv.gz
 
 if [ "$VALUE1" -a "$VALUE2" ]; then
 	echo "archa_box2_cam1;phototrapvalue;$VALUE1;$ISO" | /bin/gzip >> $PREFIX1.csv.gz
 	echo "archa_box2_cam2;phototrapvalue;$VALUE2;$ISO" | /bin/gzip >> $PREFIX2.csv.gz
 fi
 
-mv $PREFIX1.csv.gz $RAMDISK/http
-mv $PREFIX2.csv.gz $RAMDISK/http
+mv $PREFIX1.csv.gz $RAMDISK/http 2>/dev/null
+mv $PREFIX2.csv.gz $RAMDISK/http 2>/dev/null
 
 echo 'g0' >  /dev/AVR
